Added Scope, and Client
Also, retrieved the configuration using
