--- Source/IdentityServer.API/Config.cs
@@ -1,20 +1,0 @@
-﻿using Duende.IdentityServer.Models;
-
-namespace IdentityServer.API;
-
-public static class Config
-{
-    public static IEnumerable<IdentityResource> IdentityResources =>
-        new IdentityResource[]
-        { 
-            new IdentityResources.OpenId()
-        };
-
-    public static IEnumerable<ApiScope> ApiScopes =>
-        new ApiScope[]
-            { };
-
-    public static IEnumerable<Client> Clients =>
-        new Client[] 
-            { };
-}
