+++ Source/IdentityServer.API/HostingExtensions.cs
@@ -1,3 +1,4 @@
+using IdentityServer.API.Configuration;
 using Serilog;
 
 namespace IdentityServer.API;
@@ -20,11 +21,11 @@ internal static class HostingExtensions
 
         return builder.Build();
     }
-    
+
     public static WebApplication ConfigurePipeline(this WebApplication app)
-    { 
+    {
         app.UseSerilogRequestLogging();
-    
+
         if (app.Environment.IsDevelopment())
         {
             app.UseDeveloperExceptionPage();
@@ -33,7 +34,7 @@ internal static class HostingExtensions
         // uncomment if you want to add a UI
         //app.UseStaticFiles();
         //app.UseRouting();
-            
+
         app.UseIdentityServer();
 
         // uncomment if you want to add a UI
